client/server 境界整理: CompletedReportListItem を shared/types に集約
CodeRabbit 指摘: client component から server/loaders/ の型を直接 import
していたので shared/types/ に移動。loader 側はそこから re-import する形に変更。
プロンプト本文に関する nitpick (bulk-mode.ts:81-83) は本 PR スコープ外。
